refactor(sanity): move manifest extraction code
