+++ packages/sanity/src/_internal/cli/threads/extractSchema.ts
@@ -2,13 +2,9 @@ import {isMainThread, parentPort, workerData as _workerData} from 'node:worker_t
 
 import {extractSchema} from '@sanity/schema/_internal'
 import {type SchemaType} from 'groq-js'
-import {
-  ConcreteRuleClass,
-  type SchemaTypeDefinition,
-  type SchemaValidationValue,
-  type Workspace,
-} from 'sanity'
+import {type SchemaTypeDefinition, type Workspace} from 'sanity'
 
+import {extractWorkspace} from '../../manifest/extractManifest'
 import {getStudioWorkspaces} from '../util/getStudioWorkspaces'
 import {mockBrowserEnvironment} from '../util/mockBrowserEnvironment'
 
@@ -26,20 +22,8 @@ export interface ExtractSchemaWorkerData {
 type WorkspaceTransformer = (workspace: Workspace) => ExtractSchemaWorkerResult
 
 const workspaceTransformers: Record<Format, WorkspaceTransformer> = {
-  'direct': (workspace) => {
-    return {
-      name: workspace.name,
-      dataset: workspace.dataset,
-      schema: JSON.parse(
-        JSON.stringify(workspace.schema._original, (key, value) => {
-          if (key === 'validation' && isSchemaValidationValue(value)) {
-            return serializeValidation(value)
-          }
-          return value
-        }),
-      ),
-    }
-  },
+  // @ts-expect-error FIXME
+  'direct': extractWorkspace,
   'groq-type-nodes': (workspace) => ({
     schema: extractSchema(workspace.schema, {
       enforceRequiredFields: opts.enforceRequiredFields,
@@ -114,52 +98,4 @@ function getWorkspace({
 
 function isFormat(maybeFormat: string): maybeFormat is Format {
   return formats.includes(maybeFormat as Format)
-}
-
-// TODO: Simplify output format.
-function serializeValidation(validation: SchemaValidationValue): SchemaValidationValue[] {
-  const validationArray = Array.isArray(validation) ? validation : [validation]
-
-  return validationArray
-    .reduce<SchemaValidationValue[]>((output, validationValue) => {
-      if (typeof validationValue === 'function') {
-        const rule = new ConcreteRuleClass()
-        const applied = validationValue(rule)
-
-        // TODO: Deduplicate by flag.
-        // TODO: Handle merging of validation rules for array items.
-        return [...output, applied]
-      }
-      return output
-    }, [])
-    .flat()
-}
-
-function isSchemaValidationValue(
-  maybeSchemaValidationValue: unknown,
-): maybeSchemaValidationValue is SchemaValidationValue {
-  if (Array.isArray(maybeSchemaValidationValue)) {
-    return maybeSchemaValidationValue.every(isSchemaValidationValue)
-  }
-
-  // TODO: Errors with `fields() can only be called on an object type` when it encounters
-  // the `fields` validation rule on a type that is not directly an `object`. This mayb be
-  // because the validation rules aren't normalized.
-  try {
-    return (
-      maybeSchemaValidationValue === false ||
-      typeof maybeSchemaValidationValue === 'undefined' ||
-      maybeSchemaValidationValue instanceof ConcreteRuleClass ||
-      (typeof maybeSchemaValidationValue === 'function' &&
-        isSchemaValidationValue(maybeSchemaValidationValue(new ConcreteRuleClass())))
-    )
-  } catch (error) {
-    const hasMessage = 'message' in error
-
-    if (!hasMessage || error.message !== 'fields() can only be called on an object type') {
-      throw error
-    }
-  }
-
-  return false
 }
